Run as ec2-user
* On the host system, user `ec2-user` has UID 500. By creating `ec2-user`
with UID 500 in the container, we can access files shared via nfs, owned
by `ec2-user`. This is specific to amazonlinux machines.
* The `mount_nfs.sh` script needs to have root permissions to execute `rpcbind`
and the `mount` commands. This means that we need to either run the script as
root or switch to root before running that script. I opted for the latter option.
* The `/home` directory when mounted overwrites `/home/ec2-user` which is the
traditional home directrory for user `ec2-user`. Hence I changed the home directory
to `/opt/ec2-user`.

Signed-off-by: Sean Smith <seaam@amazon.com>